backend/requirements.txt
flask==3.1.0
flask-cors==5.0.0
flask-limiter==3.5.0
PyJWT==2.8.0
scikit-learn>=1.5.0
pandas>=2.2.0
numpy>=1.26.0
psycopg2-binary==2.9.11
sqlalchemy==2.0.36
python-dotenv==1.0.1
joblib==1.4.2
geopy>=2.4.0
